+++ database/schemas/all.schema.mjs
@@ -1,0 +1,17 @@
+import chatSchema from '#schemas/chat.schema';
+import messageSchema from '#schemas/message.schema.js';
+import prisonSchema from '#schemas/prison.schema.js';
+import prisonerSchema from '#schemas/prisoner.schema.js';
+import ruleSchema from '#schemas/rule.schema.js';
+import userSchema from '#schemas/user.schema.js';
+
+export default class Schemas {
+    static chat = chatSchema;
+    static message = messageSchema;
+    static prison = prisonSchema;
+    static prisoner = prisonerSchema;
+    static rule = ruleSchema;
+    static user = userSchema;
+};
+
+
