+++ database/schemas/prison.schema.js
